Delete ~/.git-credentials after pipeline clone (#20)
Delete ~/.git-credentials after pipeline clone

... to not interfere with Git credentials the pipeline uses to access other repositories on the same Git server.

`~/.git-credentials` contains an entry to use the pipeline clone secret for the Git server where the pipeline definition should be taken from. This credential must have the authorization to clone the pipeline repository, but may NOT have the authorization to access other repositories. If the pipeline needs to access other repositories on the _same_ Git server using other credentials, `~/.git-credentials` takes precedence and the wrong credentials are used, possibly leading to failed Git operations.

Co-Authored-By: Andreas Brehmer <38314220+anbrsap@users.noreply.github.com>