Introduce Steward interface for container entrypoint
Steward creates Jenkinsfile Runner pods via Tekton task runs. To work
around a problem with container registry rate limits, Steward must
specify the JFR container entry point explicitly in the container spec
so that Tekton does not need to download the image manifest.

With this change a stable container entrypoint is defined and
established (`/steward-interface/entrypoint`), which can be implemented
by the image as desired.